requirements.txt
rasa==2.4.3
rasa-x==0.38.1
python-dotenv